Separate IamEventCategory enum
