+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/IamEventCategory.java
@@ -1,0 +1,15 @@
+package it.infn.mw.iam.audit.events;
+
+public enum IamEventCategory {
+  NONE,
+  ACCOUNT,
+  GROUP,
+  REGISTRATION,
+  AUTHENTICATION,
+  AUTHORIZATION,
+  SCOPE_POLICY,
+  AUP,
+  MEMBERSHIP,
+  CLIENT,
+  TOKEN
+}
