+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/aup/AupEvent.java
@@ -16,6 +16,7 @@
 package it.infn.mw.iam.audit.events.aup;
 
 import it.infn.mw.iam.audit.events.IamAuditApplicationEvent;
+import it.infn.mw.iam.audit.events.IamEventCategory;
 import it.infn.mw.iam.persistence.model.IamAup;
 
 public abstract class AupEvent extends IamAuditApplicationEvent {
