+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/aup/AupSignedEvent.java
@@ -20,6 +20,7 @@ import static java.lang.String.format;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import it.infn.mw.iam.audit.events.IamAuditApplicationEvent;
+import it.infn.mw.iam.audit.events.IamEventCategory;
 import it.infn.mw.iam.audit.utils.IamAupSignatureSerializer;
 import it.infn.mw.iam.persistence.model.IamAupSignature;
