+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/client/ClientEvent.java
@@ -20,6 +20,7 @@ import org.mitre.oauth2.model.ClientDetailsEntity;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import it.infn.mw.iam.audit.events.IamAuditApplicationEvent;
+import it.infn.mw.iam.audit.events.IamEventCategory;
 import it.infn.mw.iam.audit.utils.IamClientSerializer;
 
 public abstract class ClientEvent extends IamAuditApplicationEvent {
