+++ iam-login-service/src/main/java/it/infn/mw/iam/audit/events/group/GroupEvent.java
@@ -18,6 +18,7 @@ package it.infn.mw.iam.audit.events.group;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import it.infn.mw.iam.audit.events.IamAuditApplicationEvent;
+import it.infn.mw.iam.audit.events.IamEventCategory;
 import it.infn.mw.iam.audit.utils.IamGroupSerializer;
 import it.infn.mw.iam.persistence.model.IamGroup;
